refactor: split advance setting
